+++ lua/react_suspense_lens/init.lua
@@ -1,3 +1,0 @@
--- Backwards-compatible module name.
--- Keep this for users/configs that still do: require('react_suspense_lens')
-return require('react-suspense-lens')